perf: optimize math.Int.Size for small values
This is a simpler version of #16263 that only optimizes values that fit
in 53 bits. It is possible to optimize values that fit in 64 bits by
using big.Int.BitLen and math.Log2(10), but it doesn't seem worth the
complexity, especially given the revert of #16263.

I failed to beat big.Int.Marshal for values that don't fit in 64 bits.